sunrpc: extract some common sunrpc_cache code from nfsd
Patchwork-id: 35346
O-Subject: [RHEL6.1 PATCH 1/5] sunrpc: extract some common sunrpc_cache code
	from nfsd (take 2)
Bugzilla: 690900
RH-Acked-by: David Howells <dhowells@redhat.com>
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

Just to keep this the same as in RHEL5

From: NeilBrown <neilb@suse.de>

Rather can duplicating this idiom twice, put it in an inline function.
This reduces the usage of 'expiry_time' out side the sunrpc/cache.c
code and thus the impact of a change that is about to be made to that
field.

Signed-off-by: Steve Dickson <steved@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>