ixgbe: balance free_irq calls with request_irq calls
Patchwork-id: 35363
O-Subject: [PATCH RHEL6.1] ixgbe: balance free_irq calls with request_irq calls
Bugzilla: 692988
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>

The current ixgbe driver was incorrectly freeing IRQs that we had not
requested.  This change corrects that by making certain we only free
q_vectors that we have requested IRQs for.

This is a backport of upstream commit:

894ff7cf0e0cf7596f9b0d3c30e32c87f8df2784 ixgbe: balance free_irq calls with request_irq calls

This will resolve Intel's request in RHBZ 692988.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>