spec: strip note and comment from ppc64's vmlinux before checksum is calculated
Patchwork-id: 35419
O-Subject: [RHEL6 PATCH] spec: strip note and comment from ppc64's vmlinux
	before checksum is calculated
Bugzilla: 692515
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=692515

After the kernel is built, sha512hmac is used to get the kernel image's
checksum and store it under /boot/.vmlinuz-<version> so FIPS can have it
checked at boot time. On ppc64, the test is failing. ppc64 uses vmlinux
and it's stripped of the comment section before calculating the
checksum. problem is, in Red Hat's RPM macros, the post install script
brp-strip-comment-note will think vmlinux is a regular executable and
will strip it from the note section too, changing the final checksum.
And that's useful, otherwise the kernel package will be 10x of the
regular size.

This patch strips both comment and note section before the checksum is
calculated.

Boot tested in a ppc64 and crash seems to work fine with the regular
kernel + debuginfo. Checksum matches.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>