mce: don't print "human readable" message for corrected errors
Patchwork-id: 35431
O-Subject: [RHEL6.1 PATCH BZ 688281]: x86: mce: don't print "human readable"
	message for corrected errors
Bugzilla: 688281
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Bob Picco <bpicco@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>

This has likely become more visible due to the issue in BZ 694891 (which is
also a blocker, and in NEEDINFO of Intel).

We've had 2 different BZs filed against this problem, and a third complaint
from a customer via email.  While this doesn't traditionally meet the
definition of a blocker, given the # of complaints and the lightweight nature
of the patch I've flipped the blocker flag for this BZ in 6.1.

P.

---8<---
Patch from Andi's MCE staging tree:

http://git.kernel.org/?p=linux/kernel/git/ak/linux-mce-2.6.git;a=commitdiff;h=6e3c7411d2b86bff210c59caa432e8e862037bfd

x86: mce: don't print "human readable" message for corrected errors

The issue here is that a corrected error does not output any information other
than a message that says

[Hardware Error]: No human readable MCE decoding support on this CPU type.
[Hardware Error]: Run the message through 'mcelog --ascii' to decode.

This is confusing because there is no message to run through 'mcelog --ascii'
for a correctable error.

Disable the message for corrected errors.

Successfully tested by me.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>