libfc: Fixing a memory leak when destroying an interface
Patchwork-id: 35433
O-Subject: [RHEL6.1 PATCH 02/15] libfc: Fixing a memory leak when destroying an
	interface
Bugzilla: 691611
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>

From: Mike Christie <mchristi@redhat.com>

>From upstream commit fe5e3f1aec310779a4b830022a26842b8d587228

From: Parikh, Neerav <neerav.parikh@intel.com>

Subject: [SCSI] libfc: Fixing a memory leak when destroying an interface

When an fcoe interface is being destroyed; in the process the
fcoe driver will try to release all the resources it had allocated
for that interface including rports. But, it seems that it does not
release the reference held for the name server rport in that process
resulting into a memory leak. This patch fixes that memory leak.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>