perf: Complain louder about BIOSen corrupting CPU/PMU state and continue
Patchwork-id: 35465
O-Subject: [RHEL6 PATCH] perf,
	x86: Complain louder about BIOSen corrupting CPU/PMU state and
	continue
Bugzilla: 694913
RH-Acked-by: Matthew Garrett <mjg@redhat.com>
RH-Acked-by: Bob Picco <bpicco@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=694913

This is a sticky point with one of our customers because currently we just
disable the perf subsystem when we detect the BIOS is using it.  Upstream isn't
very happy by doing this and prefer the BIOS vendor to stop abusing the PMU,
but it at least allows some developers to continue doing their work despite
the fact that perf events may be corrupted.

Tested on x86_64 bare metal, kdump, and kvm.  Bare metal now works on those
machines.  Kdump will work again (though it still spits out an error from
improper shutdown).  KVM is still trapped because the write/read check still
fails (which is expected).

I would like this to go into 6.1 because there is already a lot of debates
back and forth on this and the instructions the vendor provides (to disable
the BIOS use) doesn't always work correctly.

commit 45daae575e08bbf7405c5a3633e956fa364d1b4f
Author: Ingo Molnar <mingo@elte.hu>
Date:   Fri Mar 25 10:24:23 2011 +0100

    perf, x86: Complain louder about BIOSen corrupting CPU/PMU state and continue

    Eric Dumazet reported that hardware PMU events do not work on his
    system, due to the BIOS corrupting PMU state:

        Performance Events: PEBS fmt0+, Core2 events, Broken BIOS detected, using software events only.
        [Firmware Bug]: the BIOS has corrupted hw-PMU resources (MSR 186 is 43003c)

    Linus suggested that we continue in the face of such BIOS-induced CPU
    state corruption:

       http://lkml.org/lkml/2011/3/24/608

    Such BIOSes will have to be fixed - Linux developers rely on a working and
    fully capable PMU and the BIOS interfering with the CPU's PMU state is simply
    not acceptable.

    So this patch changes perf to continue when it detects such BIOS
    interaction, some hardware events may be unreliable due to the BIOS
    writing and re-writing them - there's not much the kernel can do
    about that but to detect the corruption and report it.

    Reported-and-tested-by: Eric Dumazet <eric.dumazet@gmail.com>
    Suggested-by: Linus Torvalds <torvalds@linux-foundation.org>
    Acked-by: Peter Zijlstra <a.p.zijlstra@chello.nl>
    Cc: Thomas Gleixner <tglx@linutronix.de>
    Cc: Arnaldo Carvalho de Melo <acme@redhat.com>
    Cc: Frederic Weisbecker <fweisbec@gmail.com>
    Cc: Mike Galbraith <efault@gmx.de>
    Cc: Steven Rostedt <rostedt@goodmis.org>
    LKML-Reference: <new-submission>
    Signed-off-by: Ingo Molnar <mingo@elte.hu>