Call PCIe _OSC methods earlier
Patchwork-id: 35477
O-Subject: [PATCH] [RHEL 6.1 PATCH] Call PCIe _OSC methods earlier
Bugzilla: 693974
RH-Acked-by: Stanislaw Gruszka <sgruszka@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

rhbz: 693974

In 6.0 we disabled ASPM if the PCIe _OSC methods failed to grant us full
control of the PCIe root. 6.1 reworked the way _OSC was called, but in the
process moved it from ACPI enumeration time to PCI enumeration time. This
means that PCIe roots may be enabled during pci_enable_bridges() and have
ASPM turned on, only for us to then attempt to turn it off shortly after.
Some IDT PCIe bridges appear to react violently badly to this and lock up.
This patch moves the _OSC calls to bridge enable time and caches the data
so that the method isn't called again when we come to enabling PCIe
functionality.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>