firmware: re-add the recently deleted bnx2x fw 6.2.5.0
Patchwork-id: 35480
O-Subject: [RHEL6.1 PATCH] firmware: re-add the recently deleted bnx2x fw 6.2.5.0
Bugzilla: 690470
RH-Acked-by: John Feeney <jfeeney@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=690470

Andy and Neil argued against removing firmware, even if it was not in a
released version. Let's re-add bnx2x firmware 6.2.5.0 that was deleted
recently by:
  commit aa7fe345874fd4a5a5b4fcd3ef872c7d8e974bd5:
  Author: Michal Schmidt <mschmidt@redhat.com>
  Date:   Thu Apr 7 08:43:35 2011 -0400

      [netdrv] bnx2x: Update firmware to 6.2.9
and:
  commit e67abbd651239dc1991e84ccea51991749ee6f74
  Author: Michal Schmidt <mschmidt@redhat.com>
  Date:   Thu Apr 7 15:18:34 2011 -0400

      [netdrv] firmware/bnx2x: add 6.2.9.0 fw, remove unused fw

Note that the latter commit also removed firmware 5.2.7.0 and 6.0.34.0,
but these versions were already not present in the Makefile and not
built into the kernel-firmware package.
5.2.7.0 was only used pre-6.0 GA.
6.0.34.0 came and went in the patch series leading to 2.6.32-99.el6, so
you could not ever need it if you used only official Brew builds.
Thus this patch is not re-adding these two versions.

Brew in progress:
https://brewweb.devel.redhat.com/taskinfo?taskID=3256066

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>