be2iscsi: fix be2iscsi rmmod
Patchwork-id: 35485
O-Subject: [RHEL6.1 PATCH] be2iscsi: fix be2iscsi rmmod (v2)
Bugzilla: 695585
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Rob Evers <revers@redhat.com>

From: Mike Christie <mchristi@redhat.com>

This is for bz 695585.

The patch fixes a use after free that results in a oops and a oops
that occurs if boot is not being used due to accessing a
null pointer.

The patch has been sent here:
http://marc.info/?l=linux-scsi&m=130273112001872&w=2

I have tested it and it fixes my problem here. Emulex has
also tested.

I know this is lower priority because it only occurs during
rmmod, but it is a regression from 6.0, so I would like
to get the fix in if it is ok with everyone.

v2:
Move NULL check to lib function iscsi_boot_destroy_kset instead
of having callers check.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>