igb: for 82576 EEPROMs reporting invalid size default to 16kB
Patchwork-id: 35493
O-Subject: [RHEL 6.1 PATCH] igb: for 82576 EEPROMs reporting invalid size default
	to 16kB
Bugzilla: 695751
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Michal Schmidt <mschmidt@redhat.com>
RH-Acked-by: Andy Gospodarek <gospo@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Matthew Garrett <mjg@redhat.com>

Bugzilla:
https://bugzilla.redhat.com/show_bug.cgi?id=695751

Description:
The check that gracefully handled invalid EEPROM sizes was removed by
commit 4322e561a93ec7ee034b603a6c610e7be90d4e8a. Without this check the
EEPROM validation fails if the size is invalid and the NIC is not usable
by the OS. Observed with a 8086:10c9 NIC.

igb 0000:03:00.0: 0 vfs allocated
igb 0000:03:00.0: The NVM Checksum Is Not Valid
ACPI: PCI interrupt for device 0000:03:00.0 disabled
igb: probe of 0000:03:00.0 failed with error -5

Re-introducing the check for 82576 NICs with an additional printk to
report the problem.

Upstream Status:
Not upstream yet as it was discovered very recently, but under
discussion with Intel.

Brew Build:
https://brewweb.devel.redhat.com/taskinfo?taskID=3257544

Test Status:
Tested by myself. Intel has confirmed that the EEPROM in question is
indeed reporting an invalid EEPROM size.

  Stefan

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>