nfs: use unstable writes for groups of small DIO writes
Patchwork-id: 35497
O-Subject: [RHEL6.1 PATCH] BZ#694309: nfs: use unstable writes for groups of
	small DIO writes
Bugzilla: 694309
RH-Acked-by: David Howells <dhowells@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

Currently, the client uses FILE_SYNC whenever it's writing less than or
equal data to the wsize. This is a problem though if we have a bunch
of small iovec's batched up in a single writev call. The client will
iterate over them and do a single FILE_SYNC WRITE for each.

Instead, change the code to do unstable writes when we'll need to do
multiple WRITE RPC's in order to satisfy the request. While we're at
it, optimize away the allocation of commit_data when we aren't going
to use it anyway.

I tested this with a program that allocates 256 page-sized and aligned
chunks of data into an array of iovecs, opens a file with O_DIRECT, and
then passes that into a writev call 128 times. Without this patch, it
took 5m16s to run on my (admittedly crappy) test rig. With this patch,
it finished in 7.5s.

This patch is not upstream. I've submitted it there, but apparently
the runway for 6.1 is too short for us to wait for a response from
the maintainer.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>