thp: simple fix for /dev/zero THP mprotect bug
Patchwork-id: 35539
O-Subject: [PATCH RHEL6.1] thp: simple fix for /dev/zero THP mprotect bug
Bugzilla: 690444
RH-Acked-by: Bob Picco <bpicco@redhat.com>
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Johannes Weiner <jweiner@redhat.com>

Hello,

this is a fix for bug
https://bugzilla.redhat.com/show_bug.cgi?id=690444, and it's not the
same fix that I suggested for upstream. This is a simpler and safer
approach that results in a smaller patch. It prevents THP to be used
on /dev/zero. The upstream approach allows THP to be used on /dev/zero
but I don't feel like wanting to take risks for 6.1. This patch may be
good idea to be applied to 6.0 too. I'm currently unsure about the
implications of this bug, but I believe it can lead to some vma
corruption if madvise is used on top of /dev/zero mappings, exactly
like we had in RHEL6.0 beta before fixing the page_referenced vs
referenced BUG_ON in split_huge_page.

The upstream fix (if accepted) then sounds good idea to backport for
6.2, then we can reverse this simpler patch.

===
Subject: thp: simple fix for /dev/zero THP mprotect bug

From: Andrea Arcangeli <aarcange@redhat.com>

The THP code wasn't fully enabled for vm_file backed vmas, so it's safer to
enable the THP page fault only for vmas with vm_file NULL.

Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>