GFS2 causes kernel panic in spectator mode
Patchwork-id: 35543
O-Subject: [RHEL 6.1] GFS2 causes kernel panic in spectator mode (bz #696535)
Bugzilla: 696535
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

This is a RHEL6 port of the fix which has already gone into both
upstream and RHEL5. It is fairly minimal, and has passed some tests
recently in RHEL6. The RHEL5 version was verified by the original
reporter.

Steve.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>