Adding Chelsio Firmware for cxgb4
Patchwork-id: 35548
O-Subject: [RHEL 6.1 PATCH] Adding Chelsio Firmware for cxgb4 (bz 691929)
Bugzilla: 691929
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: John Feeney <jfeeney@redhat.com>
RH-Acked-by: Jay Fenlason <fenlason@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>

From: Neil Horman <nhorman@tuxdriver.com>

Chelsio, apparently committed to trying to miss the 6.1 deadline, forgot to post
the cxgb4 firmware to the upstream firmware tree until yesterday, and IBM
discovered that makes installation pretty hard on RHEL.  This is the patch they
posted upstream yesterday afternoon.  I realize its _very_ late in the 6.1
cycle, but this is strictly just adding firmware to the build, so if theres any
chance we could sneak it in, that would be good.  Thanks

Brew:
https://brewweb.devel.redhat.com/taskinfo?taskID=3273721

Resolves bz 691929
Neil