ipr: improve interrupt service routine performance
Patchwork-id: 35607
O-Subject: [PATCH RHEL6.2 BZ696754 v2] ipr: improve interrupt service routine
	performance
Bugzilla: 696754
RH-Acked-by: David Howells <dhowells@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=696754

Description:
------------
During performance testing on P7 machines it was observed that the interrupt
service routine was doing unnecessary MMIO operations.

This patch rearranges the logic of the routine and moves some of the code out
of the main routine.  The result is that there are now fewer MMIO operations in
the performance path of the code.

As a result of the above change, an existing condition was exposed where the
driver could get an "unexpected" hrrq interrupt.  The original code would flag
the interrupt as unexpected and then reset the adapter.  After further analysis
it was confirmed that this condition can occasionally occur and that the
interrupt can safely be ignored.  Additional code in this patch detects this
condition, clears the interrupt and allows the driver to continue without
resetting the adapter.

Signed-off-by: Wayne Boyer <wayneb@linux.vnet.ibm.com>

v2- make the interrupt comment clearer

RHEL Version Found:
-------------------
RHEL 6.1

Brew:
-----
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=3277934

Upstream:
---------

posted here:
http://article.gmane.org/gmane.linux.scsi/67065

and should be accepted soon.

Aris: I'll send you the commit id once it is accepted.

Test Status:
------------
Tested by the IBM test team.

---------------------------------------------------------------
Steve Best
IBM on-site partner

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>