GFS2: make sure fallocate bytes is a multiple of blksize
Patchwork-id: 35616
O-Subject: [RHEL 6.2 PATCH] GFS2: make sure fallocate bytes is a multiple of
	blksize
Bugzilla: 695763
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

The GFS2 fallocate code chooses a target size to for allocating chunks of
space.  Whenever it can't find any resource groups with enough space free, it
halves its target. Since this target is in bytes, eventually it will no longer
be a multiple of blksize.  As long as there is more space available in the
resource group than the target, this isn't a problem, since gfs2 will use the
actual space available, which is always a multiple of blksize.  However,
when gfs couldn't fallocate a bigger chunk than the target, it was using the
non-blksize aligned number. This caused a BUG in later code that required
blksize aligned offsets.  GFS2 now ensures that bytes is always a multiple of
blksize

This fix has been submitted to Steve's upstream git tree
This fixes bz #965763

Signed-off-by: Benjamin Marzinski <bmarzins@redhat.com>