Check if PTE is already allocated during page fault
Patchwork-id: 35640
O-Subject: Re: [PATCH RHEL6.1] Check if PTE is already allocated during page
	fault
Bugzilla: 696998
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Johannes Weiner <jweiner@redhat.com>

This patch was safe and a minor improvement but I noticed in RHEL6.1
we had pte_alloc_map in memory.c already! And memory.c is the more important
place (huge_memory.c pmd_none check is less important than in memory.c
where we had it already).

Now I recall we didn't remove pte_alloc_map from RHEL6 memory.c (like
I did for upstream) for paranoid reasons (to reduce the modifications
to the code paths used when THP was off, but we preferred to leave a
race condition in case THP was disabled at runtime with sysfs).

So I now think we should probably ignore this bug for 6.1 and apply
the below for 6.2 only that will fix the race too!

====
Subject: mm: Check if PTE is already allocated during page fault

From: Mel Gorman <mgorman@suse.de>

With transparent hugepage support, handle_mm_fault() has to be careful
that a normal PMD has been established before handling a PTE fault. To
achieve this, it used __pte_alloc() directly instead of pte_alloc_map
as pte_alloc_map is unsafe to run against a huge PMD. pte_offset_map()
is called once it is known the PMD is safe.

pte_alloc_map() is smart enough to check if a PTE is already present
before calling __pte_alloc but this check was lost. As a consequence,
PTEs may be allocated unnecessarily and the page table lock taken.
Thi useless PTE does get cleaned up but it's a performance hit which
is visible in page_test from aim9.

This patch simply re-adds the check normally done by pte_alloc_map to
check if the PTE needs to be allocated before taking the page table
lock. The effect is noticable in page_test from aim9.

AIM9
                2.6.38-vanilla 2.6.38-checkptenone
creat-clo      446.10 ( 0.00%)   424.47 (-5.10%)
page_test       38.10 ( 0.00%)    42.04 ( 9.37%)
brk_test        52.45 ( 0.00%)    51.57 (-1.71%)
exec_test      382.00 ( 0.00%)   456.90 (16.39%)
fork_test       60.11 ( 0.00%)    67.79 (11.34%)
MMTests Statistics: duration
Total Elapsed Time (seconds)                611.90    612.22

(While this affects 2.6.38, it is a performance rather than a
functional bug and normally outside the rules -stable. While the big
performance differences are to a microbench, the difference in fork
and exec performance may be significant enough that -stable wants to
consider the patch)

Reported-by: Raz Ben Yehuda <raziebe@gmail.com>
Signed-off-by: Mel Gorman <mgorman@suse.de>
Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>