ibft: fix oops during boot
Patchwork-id: 35714
O-Subject: [RHEL 6.1 PATCH] ibft: fix oops during boot
Bugzilla: 698737
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>

From: Mike Christie <mchristi@redhat.com>

This is for bz 698737.

During 6.1 development we added this patch
commit ebaa80e93c565c169a2aa1647bb611416c3aec9e
Author: Mike Christie <mchristi@redhat.com>
Date:   Sat Dec 18 20:05:17 2010 -0500

    [scsi] ibft: Update iBFT handling for v1.03 of the spec

from upstream commit:
commit 4e639fdf0d0d745648aa62228ab8a0d9c03a563f
Author: Peter Jones <pjones@redhat.com>
Date:   Thu Feb 25 15:37:17 2010 -0500

    ibft: Update iBFT handling for v1.03 of the spec.

Before thet patch we just searched memory for a ibft signature.
With the patch we would parse the ACPI tables. There seems to
be a problem with this on IBM boxes. We do not know the
exact reason yet, but it is late in 6.1, so to be safe this
patch reverts the apci table parsing that causes the problem.

The patch that caused the problem was not requested by
any specific vendor or user. I had added it to stay in
sync with upstream, because I had other patches that
built on top of it.

The patch in this mail is not a exact revert of the patch
that causes the problem, because we have other patches
that are now built on top of it.

Patch has been tested by IBM. It is not upstream. The plan
is to finish debugging then fix upstream and in rhel 6.2.