GFS2: Tag all metadata with jid of last node to change it
Patchwork-id: 35822
O-Subject: [RHEL 6.2] GFS2: Tag all metadata with jid of last node to change
	it (bz #701578)
Bugzilla: 701578
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

This patch adds the jid of the last node to touch the metadata as it is
added to the journal. This is not used by GFS2 normally, but it is very
useful for debugging purposes to know which was the last node to touch
a certain block of data.

The field in the metadata header which is being used for this was unused
previously, so there are no backward compatibility issues. This patch
has been upstream for some time.

This resolves bz #701578

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>