cifs: check for private_data before trying to put it
Patchwork-id: 35830
O-Subject: [RHEL6.1 PATCH] BZ#702642: cifs: check for private_data before trying
	to put it
Bugzilla: 702642
CVE: CVE-2011-1771
RH-Acked-by: Igor Mammedov <imammedo@redhat.com>
RH-Acked-by: David Howells <dhowells@redhat.com>

Late breaking cifs patch for 6.1 for a regression. This fixes a
reproducible oops when someone tries to create a new file on cifs with
O_DIRECT|O_CREAT.

Patch is in mainline and stable.

--------------------[snip]-----------------------

cifs_close doesn't check that the filp->private_data is non-NULL before
trying to put it. That can cause an oops in certain error conditions
that can occur on open or lookup before the private_data is set.

Reported-by: Ben Greear <greearb@candelatech.com>
CC: Stable <stable@kernel.org>
Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Steve French <sfrench@us.ibm.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>