megaraid_sas: Driver only report tape drive, JBOD and logic drives
Patchwork-id: 35841
O-Subject: [RHEL6.2 PATCH] megaraid_sas:  Driver only report tape drive, JBOD
	and logic drives
Bugzilla: 619422
RH-Acked-by: Linda Wang <lwang@redhat.com>
RH-Acked-by: Mike Christie <mchristi@redhat.com>

This is for bz#619422

Upstream
http://www.spinics.net/lists/linux-scsi/msg48474.html

This patch removes a warning while boot in bz#619422 and a panic in 703087
The problem with the patch is that while it was posted upstream by LSI - and
we took it for 6.1, it was later rescinded. LSI posted instead a driver rewrite
with some files renames - that mainly to add support for new devices. This is what we
have now in 6.2. The ecstasy about this patch has also faded by LSI because
with this patch the expanders aren't longer exposed to the os.
This situation is unlucky, but this patch  can still be seen as a workaround
for the issues and at the moment we don't have another solution

Patch is tested by Fujitsu for both bz's.

Tomas