ixgbe: do not clear FCoE DDP error status for received ABTS
Patchwork-id: 35871
O-Subject: [PATCH RHEL6.2] ixgbe: do not clear FCoE DDP error status for
	received ABTS
Bugzilla: 695966
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>

Intel recently discovered a problem with FCoE with their upstream ixgbe
driver and requested that we add it as soon as possible.  This is being
posted for 6.2, so it can also go into 6.1.z.  The upstream commit is:

7aba7b077f638deb9569e0b36256cd9ae76e468c ixgbe: do not clear FCoE DDP error status for received ABTS

This will resolve Intel's request in RHBZ 695966.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>