ext3: Always set dx_node's fake_dirent explicitly
Patchwork-id: 35908
O-Subject: [PATCH RHEL6] ext3: Always set dx_node's fake_dirent explicitly.
Bugzilla: 703329
RH-Acked-by: Josef Bacik <josef@redhat.com>

This is for Bug 703329, ext3 filesystem corruption during graceful shutdown

The problem surfaced on RHEL5, and the fix is the same for RHEL6.

It's not so much corruption per se, but it's some extra zeroing
expected/required by e2fsck, to not interpret an htree as
corrupted.

It's a simple application of my upstream commit.

Thanks,
-Eric

commit d7433142b63d727b5a217c37b1a1468b116a9771
Author: Eric Sandeen <sandeen@redhat.com>
Date:   Fri Mar 4 16:04:08 2011 -0600

    ext3: Always set dx_node's fake_dirent explicitly.

    (crossport of 1f7bebb9e911d870fa8f997ddff838e82b5715ea
    by Andreas Schlick <schlick@lavabit.com>)

    When ext3_dx_add_entry() has to split an index node, it has to ensure that
    name_len of dx_node's fake_dirent is also zero, because otherwise e2fsck
    won't recognise it as an intermediate htree node and consider the htree to
    be corrupted.

    CC: stable@kernel.org
    Signed-off-by: Eric Sandeen <sandeen@redhat.com>
    Signed-off-by: Jan Kara <jack@suse.cz>