GFS2: resource group bitmap corruption resulting in panics and withdraws
Patchwork-id: 35928
O-Subject: [RHEL6.2 Patch] GFS2: resource group bitmap corruption resulting in
	panics and withdraws - bz 702057
Bugzilla: 702057
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

Hi,

This patch is for bugzilla bug #702057, which is a problem that was
identified in RHEL5 in another bug and ported.  The RHEL5 patch
was previously posted and accepted into the kernel.  This is a RHEL6
crosswrite of this upstream patch:

http://git.kernel.org/?p=linux/kernel/git/steve/gfs2-2.6-nmw.git;a=commitdiff;h=d192a8e5c6fec4fe8cdafebccc415db4074dee88

Basically, the problem was caused by a race in which an unlink or other
similar inode operation checked the number of links before the cluster
glock was held.  While waiting for the glock, another node could slip
in and change the number of links.  The solution is to re-check the
number of links after the glock has been held.

It was tested on my RHEL6 gfs-i24c-0{1,2} cluster with my reproducer program.

Regards,

Bob Peterson
Red Hat File Systems

Signed-off-by: Bob Peterson <rpeterso@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>