cio: prevent purging of CCW devices in the online state
Patchwork-id: 35992
O-Subject: [RHEL6.2 PATCH 02/14] [s390x] cio: prevent purging of CCW devices in the online state
Bugzilla: 695377

Description
-----------
After the performance of the cio_ignore purge function
(cio_ignore --purge or echo purge > /proc/cio_ignore), a device
which is in the online state and on the cio_ignore blacklist is
removed and no longer accessible.

The cio_ignore purge function should only remove devices which
are in the offline state and on the cio_ignore blacklist.  If a
purge function is performed, devices which are offline and on the
blacklist are scheduled for removal. Due to missing checks in the
set online function, such a device can still be set online.
This can result in a situation where a device is set online and
immediately afterwards removed.

When users request that a device be set online which there is
delayed work pending for this device, reject the request.  Also
ensure that no devices which are in the process of being set online
are marked for removal during a purge operation.

Bugzilla
--------
BZ 695377
https://bugzilla.redhat.com/show_bug.cgi?id=695377

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=a2fc8485f838ccd2ce5db690f81ac086489a9e7e

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>