dasd: fix race between open and offline
Patchwork-id: 35994
O-Subject: [RHEL6.2 PATCH 03/14] [s390x] dasd: fix race between open and offline
Bugzilla: 695677

Description
-----------
Oops when dasd_open tries to dereference an invalid pointer.

The dasd_open function uses the private_data pointer of
the gendisk to find the drivers internal structures that
represent this DASD device. When a DASD device is set
offline, and a process tries to open the device at the
same time, then there is a small race window, in which
dasd_open could first read a pointer from the
private_data field and then try to use it, after the
structure has already been freed.

To close this race window, store a pointer to a different
internal structure (dasd_devmap) in the private_data field.
The devmap entries are not deleted when a device is set
offline, and we already have proper locking and reference
counting in place, so that we can safely get from a devmap
to the other structures of a DASD device.

Bugzilla
--------
BZ 695677
https://bugzilla.redhat.com/show_bug.cgi?id=695677

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 2.6.39
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=65f8da475995f667af5298c644707dbd9d646ca6

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>