dasd: check sense type in device change handler
Patchwork-id: 35996
O-Subject: [RHEL6.2 PATCH 04/14] [s390x] dasd: check sense type in device change handler
Bugzilla: 700688

Description
-----------
Wrong interpretation of device state change, e.g. a
reservation is assumed to be stolen although it is not.

The 32 byte sense data that the DASD device driver
receives from a DASD device can be either in 32-bit mode
or in 24-bit compatibility mode. If this is not taken
into consideration, the sense data interpretation can be
incorrect.

Always check the sense data format in byte 27, bit 0.

Bugzilla
--------
BZ 700688
https://bugzilla.redhat.com/show_bug.cgi?id=700688

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 2.6.39
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=c7a29e56f0a3526b71e9694b403b8e2cbe9c31e5

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Kyle McMartin <kmcmarti@redhat.com>